separate information display functions from utils.py
